bugs for python 2.7

+++ src/trajectory_toolkit/Plotter.py
@@ -12,7 +12,7 @@ class Plotter:
     figureId = 0
     maxPoints = 1000;
     legendLoc = 1
-    
+
     def __init__(self, figureId, subplotDim, title='', xLabels = None, yLabels = None, maxPoints=1000):
         self.maxPoints = maxPoints
         self.subplotDim = subplotDim
@@ -21,7 +21,7 @@ class Plotter:
         else:
             self.figureId = figureId
             figure(self.figureId)
-        plt.ion()            
+        plt.ion()
         plt.show(block=False);
         plt.gcf().subplots_adjust(bottom=0.15)
         self.yLabels = yLabels
@@ -36,8 +36,8 @@ class Plotter:
                 self.xLabels.append('')
         axis = subplot(self.subplotDim[0], self.subplotDim[1], 1)
         plt.title(title)
-                
-    
+
+
     def addDataToSubplot(self, td, colID, plotID, formatstring, legend=''):
         self.colIDs.append(colID);
         self.td.append(td)
@@ -51,21 +51,21 @@ class Plotter:
         if legend != '':
             plt.legend(loc=self.legendLoc)
         self.refreshSingleLine(len(self.colIDs)-1)
-        
+
     def addDataToSubplotMultiple(self, td, col, plotID, formatstring, legend):
         colIDs = td.getColIDs(col)
         for i in xrange(0,len(colIDs)):
             self.addDataToSubplot(td,colIDs[i],plotID[i],formatstring[i],legend[i])
-        
+
     def refreshSingleLine(self,lineID):
         figure(self.figureId)
-        stepSize = floor(self.td[lineID].end()/self.maxPoints)+1;
+        stepSize = int(floor(self.td[lineID].end()/self.maxPoints))+1;
         self.lines[lineID].set_xdata(self.td[lineID].col(0)[0::stepSize])
         self.lines[lineID].set_ydata(self.td[lineID].col(self.colIDs[lineID])[0::stepSize])
         self.axes[lineID].relim()
         self.axes[lineID].autoscale_view(True,True,True)
         plt.draw()
-    
+
     def setAxis(self, plotID, t0 = None, t1 = None, y0 = None, y1 = None):
         figure(self.figureId)
         axis = subplot(self.subplotDim[0], self.subplotDim[1], plotID)
@@ -80,7 +80,7 @@ class Plotter:
             y1 = ymax
         axis.axis([t0,t1,y0,y1])
         plt.draw()
-    
+
     def setFigureSize(self,w = None,h = None):
         fig = figure(self.figureId)
         size_old = fig.get_size_inches()
@@ -90,12 +90,12 @@ class Plotter:
         if h==None:
             h = size_old[1]
         fig.set_size_inches(w, h, forward=True)
-    
+
     def removeTicks(self, plotID):
         figure(self.figureId)
         axis = subplot(self.subplotDim[0], self.subplotDim[1], plotID)
         axis.set_xticks([])
-        
+
     def refresh(self): # More efficient than refreshSingleLine
         figure(self.figureId)
         for i in xrange(0,len(self.colIDs)):
